requirements.txt
pylertalertmanager==0.1.1
